+++ src/main/java/com/dev/devblog/config/SecurityConfig.java
@@ -40,6 +40,9 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
         http.exceptionHandling()
                 .accessDeniedPage("/denied");
+
+        http.authorizeRequests().antMatchers("/kakao/callback").permitAll()
+                .anyRequest().authenticated().and().oauth2Login();
     }
