+++ src/main/java/com/dev/devblog/oauth/kakao/KakaoRepository.java
@@ -1,0 +1,9 @@
+package com.dev.devblog.oauth.kakao;
+
+import com.dev.devblog.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface KakaoRepository extends JpaRepository<User,Long> {
+    User findByUserId(String userId);
+
+}
