+++ src/main/resources/application.yml
@@ -25,7 +25,7 @@ spring:
             client_id: a3ba113686e820ae50156cdc2f58daac
             client-name: Kakao
             client_secret: CFdNPevlFNAumWBWdqyKNd6hlj1VdTlI
-            redirect-uri: "http://localhost:8080/kakaocallback"
+            redirect-uri: "http://localhost:8080/kakao/callback"
             scope: profile_nickname, account_email
 
 
@@ -43,3 +43,15 @@ spring:
 
     server:
       port: 8080
+
+kakao-info:
+  authorization-uri: https://kauth.kakao.com/oauth/authorize
+  token-uri: https://kauth.kakao.com/oauth/token
+  user-info-uri: https://kapi.kakao.com/v2/user/me
+  user-name-attribute: response
+  authorization-grant-type: authorization_code
+  client_id: a3ba113686e820ae50156cdc2f58daac
+  client-name: Kakao
+  client_secret: CFdNPevlFNAumWBWdqyKNd6hlj1VdTlI
+  redirect-uri: "http://localhost:8080/kakao/callback"
+  scope: profile_nickname, account_email
