+++ src/main/resources/templates/pages/login.html
@@ -13,7 +13,7 @@
     <input type="text" name="username" placeholder="account를 입력해주세요.">
     <input type="password" name="password" placeholder="password를 입력해주세요.">
     <button type="submit">로그인</button>
-    <a href="https://kauth.kakao.com/oauth/authorize?client_id=a3ba113686e820ae50156cdc2f58daac&redirect_uri=http://localhost:8080/kakaocallback&response_type=code"><img src="/img/kakao_login_btn.png"></a>
+    <a href="https://kauth.kakao.com/oauth/authorize?client_id=a3ba113686e820ae50156cdc2f58daac&redirect_uri=http://localhost:8080/kakao/callback&response_type=code"><img src="/img/kakao_login_btn.png"></a>
 </form>
 
 </body>
